pyproject.toml
[project]
name = "build-with-llms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate",
    "datasette>=0.65.1",
    "dotenv",
    "google-genai>=1.10.0",
    "jupyter",
    "modal>=0.73.164",
    "pymupdf>=1.25.5",
    "python-fasthtml>=0.12.12",
    "transformers",
]
